[ADD] 유저 삭제 기능 및 책 수정 기능 추가, UI 구성 변경

+++ 3차_도서관/src/Library/Library/Viewer/Admin/AdminRemoveUserViewer.cs
@@ -1,7 +1,10 @@
 using System;
 using System.Collections.Generic;
+using System.Text.RegularExpressions;
+using Library.Constants;
 using Library.Exception;
 using Library.Model;
+using Library.Utility;
 
 namespace Library
 {
@@ -11,26 +14,43 @@ namespace Library
         {
         }
 
+        public bool IsInputValid(string str, string regularExpression)
+        {
+            Regex regex = new Regex(regularExpression);
+
+            return regex.IsMatch(str);
+        }
+        
         public void RemoveUser()
         {
-            bool isEscPressed = false;
-
-            while (!isEscPressed)
+            while (true)
             {
-                KeyValuePair<bool, string> userID = inputFromUser.ReadInputFromUser(0, 0, 10, false, true);
+                Console.Clear();
+                Console.Write("Enter User ID to remove:".PadLeft(30, ' '));
+                Console.Write(new string(' ', 50));
+                KeyValuePair<bool, string> userID = inputFromUser.ReadInputFromUser(31, 0, 10, false, true);
 
                 if (!userID.Key)
                 {
-                    isEscPressed = true;
                     return;
                 }
 
-                bool success = dataManager.userManager.DeleteMember(data, Int32.Parse(userID.Value));
-
-                if (success)
+                if (IsInputValid(userID.Value, @"[0-9]+"))
                 {
-                    return;
+                    bool success = dataManager.userManager.DeleteMember(data, Int32.Parse(userID.Value));
+
+                    if (success)
+                    {
+                        Console.Clear();
+                        return;
+                    }
+                    
+                    FramePrinter.PrintOnPosition(31, 0, "유저 ID 혹은 해당 유저가 빌린 책이 있는지를 확인해주세요", AlignType.LEFT, ConsoleColor.Red);
+                    Console.ReadKey(true);
+                    Console.SetCursorPosition(31, 0);
                 }
+                
+                Console.Clear();
             }
         }
     }
